+++ src/main/java/racingcar/controller/GameController.java
@@ -1,6 +1,6 @@
 package racingcar.controller;
 
-import racingcar.domain.Car;
+import racingcar.domain.car.Car;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
