+++ src/main/java/racingcar/domain/car/DriveBehavior.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 public interface DriveBehavior {
     void drive(CarPosition carPosition);
