+++ src/main/java/racingcar/domain/car/RandomDriveBehavior.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.function.IntSupplier;
