+++ src/test/java/racingcar/domain/RacingCarsTest.java
@@ -9,6 +9,9 @@ import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import racingcar.domain.car.Car;
+import racingcar.domain.car.DriveBehavior;
+import racingcar.domain.car.RandomDriveBehavior;
 
 public class RacingCarsTest {
     @ParameterizedTest
